+++ CHANGELOG.md
@@ -17,8 +17,32 @@ Types of changes:
 
 ### Added
 - Added `QbraidJob.async_result()` to support async result retrieval using `await`. ([#945](https://github.com/qBraid/qBraid/pull/945))
+- Added `QbraidDevice.set_target_program_type`, allowing you to set a specific `ProgramSpec` (from `TargetProfile`) alias as the default ([#952](https://github.com/qBraid/qBraid/pull/952)). For example, if a device supports both "qasm2" and "qasm3", you can now restrict transpilation to one format:
+
+```python
+from qbraid.runtime import IonQProvider
+
+provider = IonQProvider()
+
+device = provider.get_device("simulator")
+
+device.metadata()["runtime_config"]["target_program_type"] # ['qasm2', 'qasm3']
+
+device.set_target_program_type("qasm2")
+
+device.metadata()["runtime_config"]["target_program_type"] # 'qasm2'
+```
+
+However the original `TargetProfile.program_spec` value remains frozen:
+
+```python
+device.profile.program_spec
+# [<ProgramSpec('builtins.str', 'qasm2')>,
+#  <ProgramSpec('builtins.str', 'qasm3')>]
+```
 
 ### Improved / Modified
+- Prepped tests for supporting `qiskit>=2.0` ([#955](https://github.com/qBraid/qBraid/pull/955))
 
 ### Deprecated
 
@@ -28,6 +52,8 @@ Types of changes:
 ### Fixed
 - Fixed Amazon Braket remote test by changing catch `JobStateError` to `TimeoutError` ([#948](https://github.com/qBraid/qBraid/pull/948))
 - Fixed upper bound of html length check in pytket circuit drawer test ([#950](https://github.com/qBraid/qBraid/pull/950))
+- Fixed simulator check for Azure target profiles ([#956](https://github.com/qBraid/qBraid/pull/956))
+- Pinned `pyqasm` version to 0.3.0 to patch CI failure due to program validation bug: https://github.com/qBraid/pyqasm/issues/175 ([#957](https://github.com/qBraid/qBraid/pull/957))
 
 ### Dependencies
 - Added `pydantic-core` to project requirements ([#946](https://github.com/qBraid/qBraid/pull/946))